requirements.txt
python-telegram-bot
selectorlib
requests
fake_useragent